requirements.txt
django==5.1
djangorestframework==3.15
django_filters==24.3